requirements.txt
contextualbandits>=0.3.14
scikit-learn
pyyaml
matplotlib
jupyter
bokeh
